requirements.txt
hydra-core>=1.3.2
hydra-submitit-launcher>=1.2.0
aim>=3.17.3
iopath>=0.1.10
torch>=2.0.0
datasets>=2.13.1
transformers>=4.31.0
xformers>=0.0.21
accelerate>=0.22.0
scikit-learn>=1.3.0
evaluate>=0.4.0
jsonlines>=4.0.0